requirements.txt
arrow==0.10.0
boto3==1.4.4
botocore==1.5.78
cobalt==0.1.9
dj-database-url==0.4.1
Django==1.10.1
django-admin-sortable2==0.6.8
django-autocomplete-light==3.2.1
django-cors-headers==1.3.1
django-filter==1.0.1
django-mptt==0.8.7
django-stdimage==2.4.1
django-storages==1.5.1
djangorestframework==3.5.3
docutils==0.13.1
futures==3.1.1
gunicorn==19.6.0
jmespath==0.9.3
lxml==3.6.4
olefile==0.44
Pillow==4.2.0
progressbar2==3.30.2
psycopg2==2.6.2
python-dateutil==2.6.0
python-utils==2.0.1
s3transfer==0.1.10
six==1.10.0
